+++ backend/src/main/java/com/celuveat/auth/query/MemberQueryService.java
@@ -1,6 +1,6 @@
 package com.celuveat.auth.query;
 
-import com.celuveat.auth.command.domain.OauthMember;
+import com.celuveat.auth.query.dao.MemberProfileResponseDao;
 import com.celuveat.auth.query.dto.MemberProfileResponse;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -9,10 +9,9 @@ import org.springframework.stereotype.Service;
 @RequiredArgsConstructor
 public class MemberQueryService {
 
-    private final OauthMemberQueryRepository oauthMemberQueryRepository;
+    private final MemberProfileResponseDao memberProfileResponseDao;
 
     public MemberProfileResponse getProfile(Long memberId) {
-        OauthMember member = oauthMemberQueryRepository.getById(memberId);
-        return new MemberProfileResponse(member.id(), member.nickname(), member.profileImageUrl());
+        return memberProfileResponseDao.find(memberId);
     }
 }
